pyproject.toml
[tool.poetry]
name = "sse-python-demo"
version = "0.1.0"
description = ""
authors = ["Anton Petrov <anton.a.petrov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.2"
fastapi = "^0.75.2"
h11 = "^0.13.0"
pydantic = "^1.9.0"
sh = "^1.14.2"
sse-starlette = "^0.10.3"
uvicorn = "^0.17.6"
aiofiles = "^0.8.0"
sseclient-py = "^1.7.2"
requests = "^2.27.1"
rich = "^12.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
